+++ R/specify.R
@@ -363,6 +363,10 @@ specify_interface_industries <- function(.tidy_iea_df,
 #'                                 Default is FALSE.
 #' @param specify_electricity_grid Boolean stating whether an electricity grid industry should be specified or not.
 #'                                 Default is FALSE.
+#' @param ascribe_eiou_to_renewable_plants A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants
+#'                                         should be ascribed to the new renewable industries. Default is FALSE.
+#' @param ascribe_eiou_to_nuclear A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants
+#'                                should be ascribed to the new nuclear industry. Default is FALSE.
 #' @param flow_aggregation_point The name of the flow aggregation point column in `.tidy_iea_df`. Default is "Flow.aggregation.point".
 #' @param eiou A string identifying energy industry own use in the flow aggregation point column. Default is "Energy industry own use".
 #' @param transformation_processes A string identifying transformation processes in the flow aggregation point column. Default is "Transformation processes".
@@ -390,6 +394,8 @@ specify_tp_eiou <- function(.tidy_iea_df,
                             route_non_specified_tp = TRUE,
                             specify_renewable_plants = FALSE,
                             specify_electricity_grid = FALSE,
+                            ascribe_eiou_to_renewable_plants = FALSE,
+                            ascribe_eiou_to_nuclear = FALSE,
                             flow_aggregation_point = "Flow.aggregation.point",
                             eiou = "Energy industry own use",
                             transformation_processes = "Transformation processes",
@@ -413,12 +419,15 @@ specify_tp_eiou <- function(.tidy_iea_df,
       specify_renewable_plants = specify_renewable_plants
     ) |> 
     split_oil_gas_extraction_eiou() %>% 
+    add_nuclear_industry(
+      ascribe_eiou_to_nuclear = ascribe_eiou_to_nuclear
+    ) %>% 
+    specify_renewable_plants(
+      specify_renewable_plants = specify_renewable_plants,
+      ascribe_eiou_to_renewable_plants = ascribe_eiou_to_renewable_plants
+    ) |> 
     route_own_use_elect_chp_heat(
       split_using_shares_of = split_own_use_elect_chp_heat_using_shares_of
-    ) |> 
-    add_nuclear_industry() %>% 
-    specify_renewable_plants(
-      specify_renewable_plants = specify_renewable_plants
     ) |> 
     route_non_specified_flows(
       route_non_specified_eiou = route_non_specified_eiou,
@@ -679,6 +688,10 @@ tp_sinks_to_nonenergy <- function(.tidy_iea_df,
 #'                                 Default is FALSE.
 #' @param specify_electricity_grid Boolean stating whether an electricity grid industry should be specified or not.
 #'                                 Default is FALSE.
+#' @param ascribe_eiou_to_renewable_plants A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants
+#'                                         should be ascribed to the new renewable industries. Default is FALSE.
+#' @param ascribe_eiou_to_nuclear A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants
+#'                                should be ascribed to the new nuclear industry. Default is FALSE.
 #'
 #' @return An enhanced and corrected version of `.tidy_iea_df` 
 #'         That is ready for physical supply-use table (PSUT) analysis.
@@ -713,7 +726,9 @@ specify_all <- function(.tidy_iea_df,
       route_non_specified_eiou = route_non_specified_eiou,
       route_non_specified_tp = route_non_specified_tp,
       specify_renewable_plants = specify_renewable_plants,
-      specify_electricity_grid = specify_electricity_grid
+      specify_electricity_grid = specify_electricity_grid,
+      ascribe_eiou_to_renewable_plants = ascribe_eiou_to_renewable_plants,
+      ascribe_eiou_to_nuclear = ascribe_eiou_to_nuclear,
     ) %>% 
     specify_bunkers() %>%
     specify_interface_industries() %>%